refactor(store): split terminal store into domain-specific slices
Split monolithic terminalStore into four focused slices for better testability and maintainability:

- terminalRegistrySlice: Terminal CRUD operations, IPC integration, and persistence
- terminalFocusSlice: Focus management and maximize state
- terminalCommandQueueSlice: Command queuing for busy agents
- terminalBulkActionsSlice: Bulk operations (close by state, restart failed agents)

Key improvements:
- Fixed focus handling bug by capturing pre-removal state in removeTerminal
- Removed double-kill in restartFailedAgents (removeTerminal already handles cleanup)
- Added validation to queueCommand to prevent orphaned commands for non-existent terminals
- Added readiness guard to processQueue to prevent race conditions
- Wired up cleanupTerminalStoreListeners in App.tsx to prevent memory leaks

Store architecture uses dependency injection to avoid circular references while maintaining separation of concerns.
