+++ src/store/terminalStore.ts
@@ -2,408 +2,97 @@
  * Terminal Store
  *
  * Zustand store for managing terminal instances and grid state.
- * Handles terminal spawning, focus management, maximize/restore, and bulk actions.
+ * This store combines multiple slices for separation of concerns:
+ *
+ * - Registry Slice: Terminal CRUD operations and process tracking
+ * - Focus Slice: Focus management and maximize state
+ * - Command Queue Slice: Command queueing for busy agents
+ * - Bulk Actions Slice: Bulk operations (close by state, restart failed)
+ *
+ * Each slice is independently testable and has a single responsibility.
  */
 
-import { create, type StateCreator } from "zustand";
-import type { TerminalInstance as TerminalInstanceType, AgentState, TerminalType } from "@/types";
+import { create } from "zustand";
+import type { AgentState } from "@/types";
+import {
+  createTerminalRegistrySlice,
+  createTerminalFocusSlice,
+  createTerminalCommandQueueSlice,
+  createTerminalBulkActionsSlice,
+  type TerminalRegistrySlice,
+  type TerminalFocusSlice,
+  type TerminalCommandQueueSlice,
+  type TerminalBulkActionsSlice,
+  type TerminalInstance,
+  type AddTerminalOptions,
+  type QueuedCommand,
+  isAgentReady,
+} from "./slices";
 
-// Re-export the shared type so callers can import from the store
-export type TerminalInstance = TerminalInstanceType;
-
-export interface AddTerminalOptions {
-  type?: TerminalType;
-  title?: string;
-  worktreeId?: string;
-  cwd: string;
-  shell?: string;
-  /** Command to execute after shell starts (e.g., 'claude' for AI agents) */
-  command?: string;
-}
+// Re-export types for consumers
+export type { TerminalInstance, AddTerminalOptions, QueuedCommand };
+export { isAgentReady };
 
 /**
- * Represents a command queued for execution when the agent becomes idle.
+ * Combined terminal store state and actions.
+ * This interface represents the full API exposed by the terminal store.
  */
-export interface QueuedCommand {
-  /** Unique identifier for this queued command */
-  id: string;
-  /** Terminal to send the command to */
-  terminalId: string;
-  /** The payload to write to the terminal */
-  payload: string;
-  /** Human-readable description (e.g., "Inject Context") */
-  description: string;
-  /** Timestamp when the command was queued */
-  queuedAt: number;
-}
+export interface TerminalGridState
+  extends TerminalRegistrySlice,
+    TerminalFocusSlice,
+    TerminalCommandQueueSlice,
+    TerminalBulkActionsSlice {}
 
-interface TerminalGridState {
-  terminals: TerminalInstance[];
-  focusedId: string | null;
-  maximizedId: string | null;
-  /** Queue of commands waiting to be sent when agents become idle */
-  commandQueue: QueuedCommand[];
+/**
+ * Create the combined terminal store.
+ *
+ * The store is composed of multiple slices, each with a single responsibility.
+ * Slices communicate through injected dependencies to avoid circular references.
+ */
+export const useTerminalStore = create<TerminalGridState>()((set, get, api) => {
+  // Helper to get terminals from the registry slice
+  const getTerminals = () => get().terminals;
+  const getTerminal = (id: string) => get().terminals.find((t) => t.id === id);
 
-  addTerminal: (options: AddTerminalOptions) => Promise<string>;
-  removeTerminal: (id: string) => void;
-  updateTitle: (id: string, newTitle: string) => void;
-  updateAgentState: (
-    id: string,
-    agentState: AgentState,
-    error?: string,
-    lastStateChange?: number
-  ) => void;
-  setFocused: (id: string | null) => void;
-  toggleMaximize: (id: string) => void;
-  focusNext: () => void;
-  focusPrevious: () => void;
+  // Create registry slice with middleware for coordinating with other slices
+  const registrySlice = createTerminalRegistrySlice({
+    onTerminalRemoved: (id, removedIndex, remainingTerminals) => {
+      // Clear command queue for this terminal
+      get().clearQueue(id);
 
-  // Command queue operations
-  /**
-   * Queue a command for a terminal. If the terminal is idle/waiting, sends immediately.
-   * If the terminal is busy (working), queues for later execution.
-   */
-  queueCommand: (terminalId: string, payload: string, description: string) => void;
-  /**
-   * Process the command queue for a terminal. Called when agent transitions to 'waiting'.
-   * Sends the first queued command (FIFO) and removes it from the queue.
-   */
-  processQueue: (terminalId: string) => void;
-  /**
-   * Clear all queued commands for a terminal (e.g., when terminal is closed).
-   */
-  clearQueue: (terminalId: string) => void;
-  /**
-   * Get the number of queued commands for a terminal.
-   */
-  getQueueCount: (terminalId: string) => number;
+      // Handle focus transfer with pre-removal index and remaining terminals
+      get().handleTerminalRemoved(id, remainingTerminals, removedIndex);
+    },
+  })(set, get, api);
 
-  // Bulk actions
-  bulkCloseByState: (states: AgentState | AgentState[]) => void;
-  bulkCloseByWorktree: (worktreeId: string, state?: AgentState) => void;
-  bulkCloseAll: () => void;
-  restartFailedAgents: () => Promise<void>;
-  getCountByState: (state: AgentState) => number;
-  getCountByWorktree: (worktreeId: string, state?: AgentState) => number;
-}
+  // Create focus slice with terminal getter
+  const focusSlice = createTerminalFocusSlice(getTerminals)(set, get, api);
 
-const TYPE_TITLES: Record<TerminalType, string> = {
-  shell: "Shell",
-  claude: "Claude",
-  gemini: "Gemini",
-  custom: "Terminal",
-};
+  // Create command queue slice with terminal getter
+  const commandQueueSlice = createTerminalCommandQueueSlice(getTerminal)(set, get, api);
 
-const createTerminalStore: StateCreator<TerminalGridState> = (set, get) => ({
-  terminals: [],
-  focusedId: null,
-  maximizedId: null,
-  commandQueue: [],
+  // Create bulk actions slice with required dependencies
+  const bulkActionsSlice = createTerminalBulkActionsSlice(
+    getTerminals,
+    (id) => get().removeTerminal(id),
+    (options) => get().addTerminal(options)
+  )(set, get, api);
 
-  addTerminal: async (options) => {
-    const type = options.type || "shell";
-    const title = options.title || TYPE_TITLES[type];
+  // Combine all slices
+  return {
+    ...registrySlice,
+    ...focusSlice,
+    ...commandQueueSlice,
+    ...bulkActionsSlice,
 
-    try {
-      // Spawn the PTY process via IPC
-      const id = await window.electron.terminal.spawn({
-        cwd: options.cwd,
-        shell: options.shell,
-        cols: 80,
-        rows: 24,
-        command: options.command,
-        type,
-        title,
-        worktreeId: options.worktreeId,
-      });
-
-      // Agent terminals (claude/gemini) start in 'idle' state
-      const isAgentTerminal = type === "claude" || type === "gemini";
-      const terminal: TerminalInstance = {
-        id,
-        type,
-        title,
-        worktreeId: options.worktreeId,
-        cwd: options.cwd,
-        cols: 80,
-        rows: 24,
-        agentState: isAgentTerminal ? "idle" : undefined,
-        lastStateChange: isAgentTerminal ? Date.now() : undefined,
-      };
-
-      set((state) => {
-        const newTerminals = [...state.terminals, terminal];
-
-        // Persist terminal list to electron-store
-        window.electron.app
-          .setState({
-            terminals: newTerminals.map((t) => ({
-              id: t.id,
-              type: t.type,
-              title: t.title,
-              cwd: t.cwd,
-              worktreeId: t.worktreeId,
-            })),
-          })
-          .catch((error) => {
-            console.error("Failed to persist terminals:", error);
-          });
-
-        return {
-          terminals: newTerminals,
-          focusedId: id,
-        };
-      });
-
+    // Override addTerminal to also set focus
+    addTerminal: async (options: AddTerminalOptions) => {
+      const id = await registrySlice.addTerminal(options);
+      set({ focusedId: id });
       return id;
-    } catch (error) {
-      console.error("Failed to spawn terminal:", error);
-      throw error;
-    }
-  },
-
-  removeTerminal: (id) => {
-    // Kill the PTY process
-    window.electron.terminal.kill(id).catch((error) => {
-      console.error("Failed to kill terminal:", error);
-      // Continue with state cleanup even if kill fails
-    });
-
-    set((state) => {
-      const newTerminals = state.terminals.filter((t) => t.id !== id);
-      const currentIndex = state.terminals.findIndex((t) => t.id === id);
-
-      // Determine new focused terminal
-      let newFocusedId: string | null = null;
-      if (state.focusedId === id && newTerminals.length > 0) {
-        // Focus the next terminal, or the previous if we removed the last one
-        const nextIndex = Math.min(currentIndex, newTerminals.length - 1);
-        newFocusedId = newTerminals[nextIndex]?.id || null;
-      } else if (state.focusedId !== id) {
-        newFocusedId = state.focusedId;
-      }
-
-      // Persist updated terminal list
-      window.electron.app
-        .setState({
-          terminals: newTerminals.map((t) => ({
-            id: t.id,
-            type: t.type,
-            title: t.title,
-            cwd: t.cwd,
-            worktreeId: t.worktreeId,
-          })),
-        })
-        .catch((error) => {
-          console.error("Failed to persist terminals:", error);
-        });
-
-      return {
-        terminals: newTerminals,
-        focusedId: newFocusedId,
-        maximizedId: state.maximizedId === id ? null : state.maximizedId,
-        // Clear any queued commands for this terminal
-        commandQueue: state.commandQueue.filter((c) => c.terminalId !== id),
-      };
-    });
-  },
-
-  setFocused: (id) => set({ focusedId: id }),
-
-  updateTitle: (id, newTitle) => {
-    set((state) => {
-      const terminal = state.terminals.find((t) => t.id === id);
-      if (!terminal) return state;
-
-      // Use trimmed title, or fall back to default title based on type
-      const effectiveTitle = newTitle.trim() || TYPE_TITLES[terminal.type];
-      const newTerminals = state.terminals.map((t) =>
-        t.id === id ? { ...t, title: effectiveTitle } : t
-      );
-
-      // Persist updated terminal list
-      window.electron.app
-        .setState({
-          terminals: newTerminals.map((t) => ({
-            id: t.id,
-            type: t.type,
-            title: t.title,
-            cwd: t.cwd,
-            worktreeId: t.worktreeId,
-          })),
-        })
-        .catch((error) => {
-          console.error("Failed to persist terminals:", error);
-        });
-
-      return { terminals: newTerminals };
-    });
-  },
-
-  updateAgentState: (id, agentState, error, lastStateChange) => {
-    set((state) => {
-      const terminal = state.terminals.find((t) => t.id === id);
-      if (!terminal) {
-        console.warn(`Cannot update agent state: terminal ${id} not found`);
-        return state;
-      }
-
-      const newTerminals = state.terminals.map((t) =>
-        t.id === id
-          ? {
-              ...t,
-              agentState,
-              error,
-              lastStateChange: lastStateChange ?? Date.now(),
-            }
-          : t
-      );
-
-      return { terminals: newTerminals };
-    });
-  },
-
-  toggleMaximize: (id) =>
-    set((state) => ({
-      maximizedId: state.maximizedId === id ? null : id,
-    })),
-
-  focusNext: () =>
-    set((state) => {
-      if (state.terminals.length === 0) return state;
-      const currentIndex = state.focusedId
-        ? state.terminals.findIndex((t) => t.id === state.focusedId)
-        : -1;
-      const nextIndex = (currentIndex + 1) % state.terminals.length;
-      return { focusedId: state.terminals[nextIndex].id };
-    }),
-
-  focusPrevious: () =>
-    set((state) => {
-      if (state.terminals.length === 0) return state;
-      const currentIndex = state.focusedId
-        ? state.terminals.findIndex((t) => t.id === state.focusedId)
-        : 0;
-      const prevIndex = currentIndex <= 0 ? state.terminals.length - 1 : currentIndex - 1;
-      return { focusedId: state.terminals[prevIndex].id };
-    }),
-
-  // Command queue operations
-  queueCommand: (terminalId, payload, description) => {
-    const terminal = get().terminals.find((t) => t.id === terminalId);
-
-    // If agent is idle/waiting, send immediately (no need to queue)
-    if (terminal?.agentState === "waiting" || terminal?.agentState === "idle") {
-      window.electron.terminal.write(terminalId, payload);
-      return;
-    }
-
-    // Otherwise, queue the command for later execution
-    const id = crypto.randomUUID();
-    set((state) => ({
-      commandQueue: [
-        ...state.commandQueue,
-        { id, terminalId, payload, description, queuedAt: Date.now() },
-      ],
-    }));
-  },
-
-  processQueue: (terminalId) => {
-    // Use functional set to avoid race conditions with concurrent queueCommand calls
-    set((state) => {
-      const forTerminal = state.commandQueue.filter((c) => c.terminalId === terminalId);
-      const remaining = state.commandQueue.filter((c) => c.terminalId !== terminalId);
-
-      // Process FIFO - send first queued command
-      if (forTerminal.length > 0) {
-        const cmd = forTerminal[0];
-        window.electron.terminal.write(cmd.terminalId, cmd.payload);
-
-        // Remove processed command, keep rest in queue
-        return { commandQueue: [...remaining, ...forTerminal.slice(1)] };
-      }
-
-      return state; // No changes if queue is empty
-    });
-  },
-
-  clearQueue: (terminalId) => {
-    set((state) => ({
-      commandQueue: state.commandQueue.filter((c) => c.terminalId !== terminalId),
-    }));
-  },
-
-  getQueueCount: (terminalId) => {
-    const { commandQueue } = get();
-    return commandQueue.filter((c) => c.terminalId === terminalId).length;
-  },
-
-  bulkCloseByState: (states) => {
-    const stateArray = Array.isArray(states) ? states : [states];
-    const { terminals, removeTerminal } = get();
-    const toRemove = terminals.filter((t) => t.agentState && stateArray.includes(t.agentState));
-    toRemove.forEach((t) => removeTerminal(t.id));
-  },
-
-  bulkCloseByWorktree: (worktreeId, state) => {
-    const { terminals, removeTerminal } = get();
-    const toRemove = terminals.filter(
-      (t) => t.worktreeId === worktreeId && (!state || t.agentState === state)
-    );
-    toRemove.forEach((t) => removeTerminal(t.id));
-  },
-
-  bulkCloseAll: () => {
-    const { terminals, removeTerminal } = get();
-    terminals.forEach((t) => removeTerminal(t.id));
-  },
-
-  restartFailedAgents: async () => {
-    const { terminals, removeTerminal, addTerminal } = get();
-    const failed = terminals.filter(
-      (t) => t.agentState === "failed" && (t.type === "claude" || t.type === "gemini")
-    );
-
-    for (const terminal of failed) {
-      try {
-        // Store config before removing
-        const config = {
-          type: terminal.type,
-          title: terminal.title,
-          worktreeId: terminal.worktreeId,
-          cwd: terminal.cwd,
-          command: terminal.type, // claude/gemini command
-        };
-
-        // Wait for terminal to be killed before respawning
-        await window.electron.terminal.kill(terminal.id);
-        removeTerminal(terminal.id);
-
-        // Small delay to ensure cleanup completes
-        await new Promise((resolve) => setTimeout(resolve, 100));
-
-        await addTerminal(config);
-      } catch (error) {
-        console.error(`Failed to restart terminal ${terminal.id}:`, error);
-        // Continue with next terminal even if one fails
-      }
-    }
-  },
-
-  getCountByState: (state) => {
-    const { terminals } = get();
-    return terminals.filter((t) => t.agentState === state).length;
-  },
-
-  getCountByWorktree: (worktreeId, state) => {
-    const { terminals } = get();
-    return terminals.filter(
-      (t) => t.worktreeId === worktreeId && (!state || t.agentState === state)
-    ).length;
-  },
+    },
+  };
 });
-
-export const useTerminalStore = create<TerminalGridState>()(createTerminalStore);
 
 // Subscribe to agent state changes from the main process
 // This runs once at module load and the cleanup function should be called on app shutdown
@@ -422,9 +111,7 @@ if (typeof window !== "undefined" && window.electron?.terminal?.onAgentStateChan
     }
 
     // Update the terminal's agent state
-    useTerminalStore
-      .getState()
-      .updateAgentState(agentId, state as AgentState, undefined, timestamp);
+    useTerminalStore.getState().updateAgentState(agentId, state as AgentState, undefined, timestamp);
 
     // Process any queued commands when agent becomes idle or waiting
     if (state === "waiting" || state === "idle") {
